// app/client/app/models/consentForm.ts
export class ConsentForm {
  consentID: string;
  userID: string;
  date: Date;
  ontarioWorks: boolean;
  owCaseWorkerName: string;
  owCaseWorkerPhone: string;
  owCaseWorkerEmail: string;
  ontarioDisabilityProgram: boolean;
  odspAgencyName: string;
  odspContactName: string;
  odspPhone: string;
  odspEmail: string;
  employmentInsurance: boolean;
  employmentServices: boolean;
  esAgencyName: string;
  esContactName: string;
  esPhone: string;
  esEmail: string;
  literacyWitness: string;
  wsib: boolean;
  wsibWtsName: string;
  wsibWtsPhone: string;
  other: string;
  contactName: string;
  contactNum: string;
  literacyAgencies: string;
}